requirements.txt
aiohttp
ffmpeg-python
beautifulsoup4

# testing purposes
requests
